python/requirements.txt
flask==2.3.3
werkzeug==2.3.7
opencv-python==4.11.0.86
pytesseract==0.3.10
numpy>=1.24.0
python-dotenv==0.19.0
requests==2.26.0
Pillow>=10.0.0
python-multipart==0.0.5
gunicorn==20.1.0
